System関連の抽出・置き換えをsystemフォルダに移動。探しづらかったため
