+++ src/features/core/extract/text/system/index.ts
@@ -1,3 +1,2 @@
 export * from "./conv";
-export * from "./system";
 export * from "./types";
