+++ src/features/core/extract/text/system/types/index.ts
@@ -2,4 +2,3 @@ export * from "./bundle";
 export * from "./item";
 export * from "./kinds";
 export * from "./meta";
-export * from "./result";
