+++ src/features/extractText.ts
@@ -56,8 +56,3 @@ export {
   extractTextFromState,
   extractTextFromWeapon,
 } from "./core/extract/text/mainData";
-
-export {
-  type ExtractedSystemTexts,
-  extractTextFromSystem,
-} from "./core/extract/text/system";
