+++ src/rmmz/system/makeSystemEx.ts
@@ -1,6 +1,13 @@
 import type { SystemTestDataSource } from "@RpgTypes/libs";
 import { makeAudioFileParams } from "@RpgTypes/libs";
-import type { System_SoundsObject } from "./core";
+import type {
+  System_SoundsObject,
+  System_TermsPartial,
+  SystemLabels_TermsParamNames,
+  Terms_Basic,
+  Terms_GameCommands,
+  Terms_Messages,
+} from "./core";
 import {
   convertTermsMessageMZtoMV,
   createMenuCommandState,
@@ -184,115 +191,127 @@ export const makeTestSystemData = ({
       gameTitle: text,
       currencyUnit: text,
     },
-    terms: {
-      basic: {
-        exp: text,
-        experience: text,
-        hp: text,
-        mp: text,
-        hpA: text,
-        mpA: text,
-        level: text,
-        levelA: text,
-        tp: text,
-        tpA: text,
-      },
-      messages: {
-        actionFailure: text,
-        actorDamage: text,
-        actorDrain: text,
-        actorGain: text,
-        actorLoss: text,
-        actorNoDamage: text,
-        actorNoHit: text,
-        actorRecovery: text,
-        alwaysDash: text,
-        autosave: text,
-        bgmVolume: text,
-        bgsVolume: text,
-        buffAdd: text,
-        buffRemove: text,
-        commandRemember: text,
-        criticalToActor: text,
-        criticalToEnemy: text,
-        counterAttack: text,
-        debuffAdd: text,
-        defeat: text,
-        escapeFailure: text,
-        emerge: text,
-        evasion: text,
-        expNext: text,
-        expTotal: text,
-        file: text,
-        enemyDamage: text,
-        enemyDrain: text,
-        enemyGain: text,
-        enemyLoss: text,
-        enemyNoDamage: text,
-        enemyNoHit: text,
-        enemyRecovery: text,
-        escapeStart: text,
-        levelUp: text,
-        possession: text,
-        saveMessage: text,
-        loadMessage: text,
-        surprise: text,
-        victory: text,
-        magicEvasion: text,
-        magicReflection: text,
-        preemptive: text,
-        meVolume: text,
-        useItem: text,
-        obtainExp: text,
-        obtainGold: text,
-        obtainItem: text,
-        obtainSkill: text,
-        partyName: text,
-        touchUI: text,
-        seVolume: text,
-        substitute: text,
-      },
-      params: {
-        agi: text,
-        atk: text,
-        def: text,
-        mat: text,
-        mdf: text,
-        luk: text,
-        eva: text,
-        hit: text,
-        mhp: text,
-        mmp: text,
-      },
-      commands: {
-        armor: text,
-        item: text,
-        skill: text,
-        equip: text,
-        status: text,
-        formation: text,
-        options: text,
-        save: text,
-        gameEnd: text,
-        cancel: text,
-        attack: text,
-        guard: text,
-        buy: text,
-        sell: text,
-        escape: text,
-        clear: text,
-        newGame: text,
-        toTitle: text,
-        continueGame: text,
-        equip2: text,
-        fight: text,
-        keyItem: text,
-        optimize: text,
-        weapon: text,
-      },
-    },
+    terms: makeTestTerms(text),
     sounds: makeSoundsEx(audio),
   });
+};
+
+export const makeTestTerms = (text: string): Required<System_TermsPartial> => {
+  const basic: Terms_Basic = {
+    exp: text,
+    experience: text,
+    hp: text,
+    mp: text,
+    hpA: text,
+    mpA: text,
+    level: text,
+    levelA: text,
+    tp: text,
+    tpA: text,
+  };
+
+  const messages: Terms_Messages = {
+    actionFailure: text,
+    actorDamage: text,
+    actorDrain: text,
+    actorGain: text,
+    actorLoss: text,
+    actorNoDamage: text,
+    actorNoHit: text,
+    actorRecovery: text,
+    alwaysDash: text,
+    autosave: text,
+    bgmVolume: text,
+    bgsVolume: text,
+    buffAdd: text,
+    buffRemove: text,
+    commandRemember: text,
+    criticalToActor: text,
+    criticalToEnemy: text,
+    counterAttack: text,
+    debuffAdd: text,
+    defeat: text,
+    escapeFailure: text,
+    emerge: text,
+    evasion: text,
+    expNext: text,
+    expTotal: text,
+    file: text,
+    enemyDamage: text,
+    enemyDrain: text,
+    enemyGain: text,
+    enemyLoss: text,
+    enemyNoDamage: text,
+    enemyNoHit: text,
+    enemyRecovery: text,
+    escapeStart: text,
+    levelUp: text,
+    possession: text,
+    saveMessage: text,
+    loadMessage: text,
+    surprise: text,
+    victory: text,
+    magicEvasion: text,
+    magicReflection: text,
+    preemptive: text,
+    meVolume: text,
+    useItem: text,
+    obtainExp: text,
+    obtainGold: text,
+    obtainItem: text,
+    obtainSkill: text,
+    partyName: text,
+    touchUI: text,
+    seVolume: text,
+    substitute: text,
+  };
+
+  const params: SystemLabels_TermsParamNames = {
+    agi: text,
+    atk: text,
+    def: text,
+    mat: text,
+    mdf: text,
+    luk: text,
+    eva: text,
+    hit: text,
+    mhp: text,
+    mmp: text,
+  };
+
+  const commands: Terms_GameCommands = {
+    armor: text,
+    item: text,
+    skill: text,
+    equip: text,
+    status: text,
+    formation: text,
+    options: text,
+    save: text,
+    gameEnd: text,
+    cancel: text,
+    attack: text,
+    guard: text,
+    buy: text,
+    sell: text,
+    escape: text,
+    clear: text,
+    newGame: text,
+    toTitle: text,
+    continueGame: text,
+    equip2: text,
+    fight: text,
+    keyItem: text,
+    optimize: text,
+    weapon: text,
+  };
+
+  return {
+    basic: basic,
+    messages: messages,
+    params: params,
+    commands: commands,
+  };
 };
 
 const makeSoundsEx = (audio: string): System_SoundsObject => {
